code reformating and added tqdm
