+++ .gitignore
@@ -164,3 +164,6 @@ snapshot/*
 output/*
 temp/*
 **/*.env
+
+# Ignore datasets
+*.tar
