+++ augmix_refactored/tools/tests.py
@@ -1,0 +1,46 @@
+import numpy as np
+import torch
+import torch.backends.cudnn as cudnn
+import torch.nn.functional as F
+
+from augmix_refactored.config import Config
+
+def test(net, test_loader, logging):
+    """Evaluate network on given dataset."""
+    net.eval()
+    total_loss = 0.
+    total_correct = 0
+    with torch.no_grad():
+        for images, targets in test_loader:
+            images, targets = images.cuda(), targets.cuda()
+            logits = net(images)
+            loss = F.cross_entropy(logits, targets)
+            pred = logits.data.max(1)[1]
+            total_loss += float(loss.data)
+            total_correct += pred.eq(targets.data).sum().item()
+
+    return total_loss / len(test_loader.dataset), total_correct / len(
+        test_loader.dataset)
+
+
+def test_c(net, test_data, base_path, config: Config, logging):
+    """Evaluate network on given corrupted dataset."""
+    corruption_accs = []
+    for corruption in CORRUPTIONS:
+        # Reference to original data is mutated
+        test_data.data = np.load(base_path + corruption + '.npy')
+        test_data.targets = torch.LongTensor(np.load(base_path + 'labels.npy'))
+
+        test_loader = torch.utils.data.DataLoader(
+            test_data,
+            batch_size=config.eval_batch_size,
+            shuffle=False,
+            num_workers=config.num_workers,
+            pin_memory=True)
+
+        test_loss, test_acc = test(net, test_loader)
+        corruption_accs.append(test_acc)
+        logging.info('{}\n\tTest Loss {:.3f} | Test Error {:.3f}'.format(
+            corruption, test_loss, 100 - 100. * test_acc))
+
+    return np.mean(corruption_accs)
