+++ snapshots/cifar10/wrn/2023-02-13_15:54:42.825759/training_log.csv
@@ -1,0 +1,55 @@
+epoch,time(s),train_loss,test_loss,test_error(%)
+001,00057,1.254900,0.00114,40.75
+002,00056,0.986349,0.00089,31.27
+003,00056,0.871266,0.00075,25.49
+004,00056,0.839802,0.00070,24.62
+005,00056,0.753722,0.00067,23.02
+006,00056,0.757237,0.00066,21.70
+007,00056,0.699703,0.00057,18.65
+008,00056,0.693307,0.00053,17.60
+009,00056,0.649432,0.00057,19.84
+010,00056,0.698830,0.00067,22.99
+011,00056,0.667225,0.00054,18.47
+012,00056,0.656592,0.00053,17.78
+013,00056,0.603903,0.00048,15.78
+014,00056,0.643583,0.00055,18.95
+015,00056,0.579404,0.00043,14.58
+016,00056,0.604459,0.00050,16.83
+017,00056,0.578136,0.00062,21.23
+018,00055,0.600708,0.00047,15.86
+019,00056,0.623234,0.00057,20.02
+020,00056,0.568289,0.00047,16.11
+021,00056,0.549508,0.00045,15.03
+022,00056,0.593671,0.00045,15.46
+023,00056,0.565958,0.00043,14.56
+024,00055,0.555969,0.00048,16.05
+025,00056,0.541564,0.00052,17.81
+026,00056,0.543121,0.00040,13.72
+027,00056,0.517380,0.00041,13.14
+028,00056,0.552985,0.00040,13.37
+029,00055,0.507505,0.00053,17.93
+030,00056,0.493713,0.00041,14.15
+031,00056,0.486144,0.00041,13.90
+032,00056,0.537307,0.00043,14.78
+033,00055,0.476358,0.00038,13.10
+034,00056,0.539758,0.00041,13.72
+035,00056,0.520866,0.00040,13.56
+036,00056,0.503530,0.00038,12.96
+037,00056,0.506263,0.00036,12.20
+038,00056,0.517747,0.00038,13.17
+039,00056,0.484241,0.00039,12.86
+040,00055,0.491141,0.00040,13.52
+041,00056,0.494967,0.00039,13.20
+042,00056,0.500599,0.00037,12.01
+043,00056,0.504017,0.00044,15.11
+044,00056,0.489071,0.00034,11.69
+045,00056,0.452539,0.00041,13.80
+046,00056,0.475546,0.00036,12.23
+047,00056,0.449602,0.00032,10.71
+048,00056,0.453485,0.00034,11.82
+049,00056,0.489894,0.00036,12.17
+050,00056,0.445577,0.00034,11.65
+051,00056,0.423919,0.00030,10.00
+052,00056,0.385044,0.00030,10.46
+053,00056,0.421280,0.00034,11.75
+054,00056,0.417573,0.00033,10.84
