+++ snapshots/cifar10/wrn/2023-02-13_16:52:10.817080/training_log.csv
@@ -1,0 +1,1 @@
+epoch,time(s),train_loss,test_loss,test_error(%)
