+++ snapshots/cifar10/wrn/2023-02-13_16:54:25.363162/training_log.csv
@@ -1,0 +1,9 @@
+epoch,time(s),train_loss,test_loss,test_error(%)
+001,00057,1.261889,0.00106,37.71
+002,00056,0.965257,0.00085,29.78
+003,00056,0.867285,0.00070,23.81
+004,00056,0.828889,0.00075,26.32
+005,00056,0.764721,0.00062,21.13
+006,00056,0.767184,0.00067,22.53
+007,00056,0.715473,0.00059,19.82
+008,00056,0.683171,0.00053,17.30
