+++ snapshots/cifar10/wrn/2023-02-13_17:02:29.419964/training_log.csv
@@ -1,0 +1,1 @@
+epoch,time(s),train_loss,test_loss,test_error(%)
