+++ snapshots/cifar10/wrn/2023-02-13_17:03:24.093972/training_log.csv
@@ -1,0 +1,2 @@
+epoch,time(s),train_loss,test_loss,test_error(%)
+001,00057,1.248842,0.00124,44.55
