+++ snapshots/cifar10/wrn/2023-02-13_17:12:59.538589/training_log.csv
@@ -1,0 +1,4 @@
+epoch,time(s),train_loss,test_loss,test_error(%)
+001,00057,1.242902,0.00118,42.95
+002,00056,0.978130,0.00084,29.44
+003,00056,0.856244,0.00071,23.90
